+++ tbot/builtin/git_tasks.py
@@ -138,7 +138,7 @@ find {patchdir} -name '*.patch'""",
         ret, _ = tb.shell.exec(f"cd {gitdir}; git am -3 {patch}", log_show_stdout=False)
         if ret != 0:  # Patch failed
             tb.shell.exec0(f"cd {gitdir}; git am --abort", log_show_stdout=False)
-            raise Exception(f"Failed to apply patch {patch}")
+            raise tbot.TestcaseFailure(f"Failed to apply patch {patch}")
 
         patchfile = tb.shell.exec0(f"cat {patch}", log_show=False)
         tbot.log.doc_appendix(
@@ -175,10 +175,9 @@ def git_bisect(
     if params is None:
         params = dict()
 
-    if and_then is None:
-        raise Exception(
-            "No test for deciding whether a commit is good or bad was provided"
-        )
+    assert (
+        and_then is not None
+    ), "No test for deciding whether a commit is good or bad was provided"
 
     assert isinstance(gitdir, pathlib.PurePosixPath)
     assert isinstance(good, str)