+++ tbot/machine/board_linux.py
@@ -82,7 +82,9 @@ class MachineBoardLinux(board.MachineBoard):
             self.own_ub = True
 
         if not isinstance(self.ub_machine, tbot.machine.MachineBoardUBoot):
-            raise Exception("Can't switch to linux if U-Boot is not yet present")
+            raise tbot.InvalidUsageException(
+                "Can't switch to linux if U-Boot is not yet present"
+            )
 
         self.powerup = False
         super()._setup(tb, previous)
@@ -95,7 +97,9 @@ class MachineBoardLinux(board.MachineBoard):
 
         self.channel = self.ub_machine.channel
         if self.channel is None:
-            raise Exception("U-Boot machine is not initialized correctly")
+            raise tbot.InvalidUsageException(
+                "U-Boot machine is not initialized correctly"
+            )
 
         self.boot_command = self.boot_command or tb.config["linux.boot_command"]
         self.login_prompt = (